@@ -1,7 +1,7 @@
 package br.edu.iesp.resource;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
 
 @Path("/aluno")
 public class AlunoResource {
