@@ -17,110 +17,60 @@
  * under the License.
  */
 
-import java.net.*;
-import java.io.*;
-import java.nio.channels.*;
-import java.util.Properties;
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.Authenticator;
+import java.net.PasswordAuthentication;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 
-public class MavenWrapperDownloader
+public final class MavenWrapperDownloader
 {
-    private static final String WRAPPER_VERSION = "3.1.1";
+    private static final String WRAPPER_VERSION = "3.2.0";
 
-    /**
-     * Default URL to download the maven-wrapper.jar from, if no 'downloadUrl' is provided.
-     */
-    private static final String DEFAULT_DOWNLOAD_URL =
-        "https://repo.maven.apache.org/maven2/org/apache/maven/wrapper/maven-wrapper/" + WRAPPER_VERSION
-            + "/maven-wrapper-" + WRAPPER_VERSION + ".jar";
+    private static final boolean VERBOSE = Boolean.parseBoolean( System.getenv( "MVNW_VERBOSE" ) );
 
-    /**
-     * Path to the maven-wrapper.properties file, which might contain a downloadUrl property to use instead of the
-     * default one.
-     */
-    private static final String MAVEN_WRAPPER_PROPERTIES_PATH = ".mvn/wrapper/maven-wrapper.properties";
-
-    /**
-     * Path where the maven-wrapper.jar will be saved to.
-     */
-    private static final String MAVEN_WRAPPER_JAR_PATH = ".mvn/wrapper/maven-wrapper.jar";
-
-    /**
-     * Name of the property which should be used to override the default download url for the wrapper.
-     */
-    private static final String PROPERTY_NAME_WRAPPER_URL = "wrapperUrl";
-
-    public static void main( String args[] )
+    public static void main( String[] args )
     {
-        System.out.println( "- Downloader started" );
-        File baseDirectory = new File( args[0] );
-        System.out.println( "- Using base directory: " + baseDirectory.getAbsolutePath() );
+        log( "Apache Maven Wrapper Downloader " + WRAPPER_VERSION );
 
-        // If the maven-wrapper.properties exists, read it and check if it contains a custom
-        // wrapperUrl parameter.
-        File mavenWrapperPropertyFile = new File( baseDirectory, MAVEN_WRAPPER_PROPERTIES_PATH );
-        String url = DEFAULT_DOWNLOAD_URL;
-        if ( mavenWrapperPropertyFile.exists() )
+        if ( args.length != 2 )
         {
-            FileInputStream mavenWrapperPropertyFileInputStream = null;
-            try
-            {
-                mavenWrapperPropertyFileInputStream = new FileInputStream( mavenWrapperPropertyFile );
-                Properties mavenWrapperProperties = new Properties();
-                mavenWrapperProperties.load( mavenWrapperPropertyFileInputStream );
-                url = mavenWrapperProperties.getProperty( PROPERTY_NAME_WRAPPER_URL, url );
-            }
-            catch ( IOException e )
-            {
-                System.out.println( "- ERROR loading '" + MAVEN_WRAPPER_PROPERTIES_PATH + "'" );
-            }
-            finally
-            {
-                try
-                {
-                    if ( mavenWrapperPropertyFileInputStream != null )
-                    {
-                        mavenWrapperPropertyFileInputStream.close();
-                    }
-                }
-                catch ( IOException e )
-                {
-                    // Ignore ...
-                }
-            }
+            System.err.println( " - ERROR wrapperUrl or wrapperJarPath parameter missing" );
+            System.exit( 1 );
         }
-        System.out.println( "- Downloading from: " + url );
 
-        File outputFile = new File( baseDirectory.getAbsolutePath(), MAVEN_WRAPPER_JAR_PATH );
-        if ( !outputFile.getParentFile().exists() )
-        {
-            if ( !outputFile.getParentFile().mkdirs() )
-            {
-                System.out.println( "- ERROR creating output directory '" + outputFile.getParentFile().getAbsolutePath()
-                    + "'" );
-            }
-        }
-        System.out.println( "- Downloading to: " + outputFile.getAbsolutePath() );
         try
         {
-            downloadFileFromURL( url, outputFile );
-            System.out.println( "Done" );
-            System.exit( 0 );
+            log( " - Downloader started" );
+            final URL wrapperUrl = new URL( args[0] );
+            final String jarPath = args[1].replace( "..", "" ); // Sanitize path
+            final Path wrapperJarPath = Paths.get( jarPath ).toAbsolutePath().normalize();
+            downloadFileFromURL( wrapperUrl, wrapperJarPath );
+            log( "Done" );
         }
-        catch ( Throwable e )
+        catch ( IOException e )
         {
-            System.out.println( "- Error downloading" );
-            e.printStackTrace();
+            System.err.println( "- Error downloading: " + e.getMessage() );
+            if ( VERBOSE )
+            {
+                e.printStackTrace();
+            }
             System.exit( 1 );
         }
     }
 
-    private static void downloadFileFromURL( String urlString, File destination )
-        throws Exception
+    private static void downloadFileFromURL( URL wrapperUrl, Path wrapperJarPath )
+        throws IOException
     {
+        log( " - Downloading to: " + wrapperJarPath );
         if ( System.getenv( "MVNW_USERNAME" ) != null && System.getenv( "MVNW_PASSWORD" ) != null )
         {
-            String username = System.getenv( "MVNW_USERNAME" );
-            char[] password = System.getenv( "MVNW_PASSWORD" ).toCharArray();
+            final String username = System.getenv( "MVNW_USERNAME" );
+            final char[] password = System.getenv( "MVNW_PASSWORD" ).toCharArray();
             Authenticator.setDefault( new Authenticator()
             {
                 @Override
@@ -130,13 +80,19 @@ public class MavenWrapperDownloader
                 }
             } );
         }
-        URL website = new URL( urlString );
-        ReadableByteChannel rbc;
-        rbc = Channels.newChannel( website.openStream() );
-        FileOutputStream fos = new FileOutputStream( destination );
-        fos.getChannel().transferFrom( rbc, 0, Long.MAX_VALUE );
-        fos.close();
-        rbc.close();
+        try ( InputStream inStream = wrapperUrl.openStream() )
+        {
+            Files.copy( inStream, wrapperJarPath, StandardCopyOption.REPLACE_EXISTING );
+        }
+        log( " - Downloader complete" );
+    }
+
+    private static void log( String msg )
+    {
+        if ( VERBOSE )
+        {
+            System.out.println( msg );
+        }
     }
 
 }
